added separate file for stories
