+++ stories/index.stories.js
@@ -6,7 +6,7 @@ import { linkTo } from '@storybook/addon-links';
 
 import { Button, Welcome } from '@storybook/react/demo';
 
-import {PropertySummary} from '../src/client/components/property-summary';
+
 storiesOf('Welcome', module).add('to Storybook', () => <Welcome showApp={linkTo('Button')} />);
 
 storiesOf('Button', module)
@@ -19,15 +19,3 @@ storiesOf('Button', module)
         </Button>
     ));
 
-let propertyItemData = {
-
-    name: 'DREAM Hostel Lviv',
-
-    location: 'Prospekt Gagarina 145, Kharkov, 61124, Ukraine ',
-
-};
-
-
-
-storiesOf('Property Summary', module)
-    .add('Item base state', () => <PropertySummary propertyItemData={propertyItemData}/>);
